Fix that disabled QRC files were enabled by QRC expansion
The automatic expansion of QRC files to the tree of their contents
didn't preserve the "enabled" property of the original QRC node.
For CMake projects that resulted in the file system tree still being
shown, with only the QRC nodes, when "Hide Disabled Files" was on,
instead of hiding the file system tree completely.

Change-Id: If1eb88478955dc8635def844cfc501befcbfc329
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>