QtSupport: Expand path variables only to local paths
It's what it used to do and actually better for the docker case where
the two only active use of these in form of %{Qt:QT_INSTALL_PREFIX}
are in a cmake build context that would require (device-)local path.

Long term we might want to use a variation of registerFileVariable
here without changing semantics on the user side.

Change-Id: If8dd77ac0b94ede41dbfe322802d5ef6c0b043ee
Reviewed-by: Christian Stenger <christian.stenger@qt.io>