CMakePM: Keep always created run configurations up to date
Qt Creator will do an update of run configurations after a project
configuration.

If an always created target was no longer existing, it shouldn't be
part of the run configuration.

For some reason this was not the case for CMake projects.

With this patchset if a target's name is changed, the previous target
name is no longer part of the run configuration.

Fixes: QTCREATORBUG-25906
Fixes: QTCREATORBUG-24914
Change-Id: I086a2540eaad9039e41fb48194d5901c7be22be8
Reviewed-by: Eike Ziller <eike.ziller@qt.io>