Update Find/Replace action even in edge cases
The Issues pane has a placeholder for the find tool bar, there is one in
the top level layout for all output panes, but it doesn't support
searching itself. Make sure that the Find/Replace action is updated even
when switching focus from another find tool bar to the issues pane.

Fixes: QTCREATORBUG-25991
Change-Id: If96e9a2c59e159408aa881a01a1df0bb7bf28963
Reviewed-by: David Schulz <david.schulz@qt.io>