EditorManager: Update current editor prior to deleting other editors
Otherwise it may happen that destructor of an editor being
deleted may trigger a call to EditorManager::currentEditor()
before reaching the editor's QObject destructor. In this case
the pointer to the editor (EditorManagerPrivate::m_currentEditor)
isn't cleared yet by Qt and it may return the pointer to
partially destructed object.

Change-Id: I4f44dce8d680008148a73eb2a8ac13477d0fb825
Reviewed-by: Eike Ziller <eike.ziller@qt.io>