QmlDesigner: Add info box when dragging material under imported 3D node
Materials of imported 3D components are not exposed outside the
component, so the materials cannot be modified by dragging a material
under the component node. Instead materials must be modified inside
the imported component itself.

Added an info box to inform users of this if they attempt to drag a
material under an imported 3D component. The info box also provides
a button to directly go into the component in question.

Fixes: QDS-4691
Change-Id: I6d859fd4716a0c23a465a91c3067534521a910aa
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>