OutputWindow: Update auto-scroll flag when "Find" changes position
In this case, the text edit's scroll bar does not emit a signal, nor is
scrollContentsBy() called (Qt bug?).
Therefore, we now check the scroll bar position also on a selection
change.
Note that this will still misbehave for the edge case where you search
for something, then scroll back to the bottom, then search for the same
string again and find the same result, because in that case the selection
does not change. We live with that.

Fixes: QTCREATORBUG-26100
Change-Id: Ib5555d9119117dd95c68175c4094bb0fd904f6e5
Reviewed-by: Orgad Shaneh <orgads@gmail.com>