Fix handling of aab checkbox in apk build step
Loosen the condition for enabling the aab builds and make it available
for any Qt version from 5.14 and above, since Google now require apps
to be packaged as aab to submit to the store, and Qt 6 always has an aab
target anyways. The multi-abi is coming for Qt 6.3 though, but aab
builds could still be allowed now.

Change-Id: If7c7b40c2d01cfe9c2887dc5aa17af48a37af43a
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>