QmlDesigner: Edit ChooseTexturePropertyDialog to accept any type
When creating a new component by dropping it to a component in the
navigator, check the parent if it contains properties that are the
same or sub type as the new created component. Those properties are
shown in a list that the created component can be binded to.

This edits the dialog that was for textures only and uses it for any
component on creation.

Task-number: QDS-4969
Change-Id: I3f1ecc6a82254f77a9bb170406ef79740359a889
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>