CMake build: Define common output directory for qt_add_qml_module
In order to enable tools like qmllint and in order to avoid the
following warning, define QT_QML_OUTPUT_DIRECTORY

Warning message:
The Tracing target is a QML module with target path QtCreator/FooBar.
It uses an OUTPUT_DIRECTORY of foobar, which should end in the same
target path, but doesn't. Tooling such as qmllint may not work
correctly.

Since this change is relevant for the build against Qt 6.2-Beta4,
common_environment.yaml gets bumped, accordingly.

Change-Id: Ic6766cf29baddd7c22b68270a0f1702aade7c53f
Reviewed-by: Eike Ziller <eike.ziller@qt.io>