auto-setup.cmake: Better handling of conan edgecases
Do not try to run the conan executable if conan executable is
not found.

If the conan auto-setup fails warn about it and provide the message
how to skip the auto-setup configuration and do not fail due to the
missing conan_paths.cmake file.

Also print a status message that informs the user that the package
manager auto-setup is performed, with the information how to skip
the step.

Task-number: QTCREATORBUG-25680
Fixes: QTCREATORBUG-25891
Fixes: QTCREATORBUG-25818
Change-Id: I0969ba9b417448f89efd22d7c361e3a54e552ed1
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Alp Öz <aoz@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>