Fix QtProjectImporter test for Qt 6.2 versions in PATH
In Qt 5, qmake is statically linked to the bootstrap library,
so just copying qmake somewhere results in a runnable executable.

In Qt 6 this changed, and qmake depends on the dynamic QtCore
library, and on Linux also on ICU. Copy these in addition to
the qmake executable. Keep the directory structure as in Qt, so
the dynamic linker finds them.

Change-Id: Ie4f6b617bc91d847b1e4fe38dc8a1c603eae4087
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>